fix: prettier install

+++ src/Commands/Install.php
@@ -489,7 +489,9 @@ SCOUT_DRIVER=database
         unlink(base_path('webpack.mix.js'));
         file_put_contents(base_path('webpack.mix.js'), file_get_contents($base.'/webpack.mix.js'));
 
-        unlink(base_path('.prettier'));
+        if (file_exists(base_path('.prettier'))) {
+            unlink(base_path('.prettier'));
+        }
         file_put_contents(base_path('.prettier'), file_get_contents($base.'/.prettier'));
     }
